+++ jdbc/src/Repository.kt
@@ -48,9 +48,11 @@ abstract class BaseCrudRepository<E: BaseEntity<ID>, ID>(db: DataSource, table: 
   protected open fun E.persister() = toValues()
 
   open fun get(id: ID, forUpdate: Boolean = false): E = db.select(selectFrom, id, "$table.id", if (forUpdate) "for update" else "") { mapper() }
-  open fun list(vararg where: PropValue<E>?, order: String = defaultOrder): List<E> =
-    db.select(selectFrom, where.filterNotNull(), order) { mapper() } // TODO: rename order to suffix
-  open fun by(vararg where: PropValue<E>?): E? = list(*where).firstOrNull() // TODO: add suffix
+
+  open fun list(vararg where: PropValue<E>?, suffix: String = defaultOrder): List<E> =
+    db.select(selectFrom, where.filterNotNull(), suffix) { mapper() }
+
+  open fun by(vararg where: PropValue<E>?, suffix: String = ""): E? = list(*where, suffix = "$suffix limit 1").firstOrNull()
   open fun count(vararg where: PropValue<E>?): Long = db.count(selectFrom, where.filterNotNull())
 
   open fun save(entity: E): Int {