+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * server: errors.on<SomeException>(StatusCode) convenience reified function added
+* jdbc: BaseCrudRepository.list() and by() now both have the suffix parameter
 
 # 1.6.8
 * server: ErrorHandler now supports finding of handlers using exception super classes